requirements.txt
requests
openai
python-dotenv
python-trading212-alinalihassan
pydantic